+++ SpriteManager.hpp
@@ -16,7 +16,7 @@ class SpriteManager {
 		~SpriteManager();
 
 		// z here is the z-index: it cannot be changed.
-		std::shared_ptr<Sprite> acquire(size_t z);
+		std::shared_ptr<Sprite> acquire(int z);
 		std::shared_ptr<Sprite> from_anim(Animation::Animation anim, bool loop);
 
 		struct Vertex {
